split get config file out
